fixed Student POJO
